requirements.txt
# 这是项目的依赖清单，请在您的Python环境中运行 pip install -r requirements.txt 来安装所有必需的库。

torch>=2.0.0
torch-geometric>=2.4.0
scanpy>=1.9.0
anndata>=0.8.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.2.0
matplotlib>=3.6.0
PyYAML>=6.0
tqdm>=4.64.0
